requirements.txt
# local package
-e .

# external requirements
numpy
opencv-python
torch
torchvision
click
Sphinx
coverage
scipy
flake8
python-dotenv>=0.5.1
